replace streaming rate configuration by burst size and timeout
--

Reason: Throttling by a fixed streaming rate is not necessary
in the presence of back-pressure.

Streaming actor timeout is now configurable by the StreamModifiedEntity
command. It allows a generous timeout to be set for policy tag streaming,
where the number of database accesses is unbounded for each policy tag
message in the stream, because a single Policy may be shared by an
unbounded amount of Things.

Sleeping time in DefaultStreamSupervisorTest is reduced.

Signed-off-by: Cai Yufei <Yufei.Cai@bosch-si.com>