add own deleted flag for search documents to increase performance for all queries that check for deletion state of a thing
-----------------------------------------

this is needed for performance reasons since a find or count with 'null' (e.g. {__deleted:null}) will cause MongoDB to go through all documents, even though an index exists. This is because it has to decide if the field exists and equals null, or doesn't exist at all (or is an empty array). Using an always filled field (__deletedFlag), MongoDB can answer the queries from the index and doesn't need to scan through all documents.

Signed-off-by: Florian Fendt <Florian.Fendt@bosch-si.com>

Signed-off-by: Florian Fendt <Florian.Fendt@bosch-si.com>