Increment cleaned-up events by 1, delete unused members of MongoReadJournal.
This removes the event with the same revision number as the
snapshot. It prevents cleanup coordinator from waking up
persistence actors for whom cleanup actions are unnecessary.

This breaks ReconnectActor - cleaned-up connections are removed
from "current persistence IDs" reported by the persistence
plugin regardless whether the target state of the connection
is "on" or not.

Signed-off-by: Cai Yufei (INST/ECS1) <yufei.cai@bosch-si.com>