Add docs about the openid connect feature
The overall documentation is for once extracted from the blog post
about OIDC and verfified with an keycloak and keycloak-gatekeeper setup.

https://www.eclipse.org/ditto/2019-08-28-openid-connect.html

For the user / operator of ditto to know how the subject is addressed
and that OIDC is a valid option in the context of authorization, an
additional bullet-point is added to the list on basic-policy. A list of
certified providers shall aid at this point.

To avoid confusion about the extend of the OIDC implementation, this
adds a notice that tokens have to be issued beforehand and
keycloak-gatekeeper is used as an example application. This is because
keycloak is a common and certified provider and gatekeeper is
maintained by the same team.

Signed-off-by: Alexander Wellbrock <a.wellbrock@mailbox.org>

Update token issue notice in the operator docs

This won't be part of ditto any time soon (if ever).

Co-Authored-By: Johannes Schneider <johannes.schneider@bosch-si.com>
Fix typo in keycloak md link

Co-Authored-By: Johannes Schneider <johannes.schneider@bosch-si.com>
Signed-off-by: Alexander Wellbrock <a.wellbrock@mailbox.org>