Test multiple client actors; disregard ack from remotely deployed client actors.
Reason: Remotely deployed actors are not started right away when their router
starts. Given how much error handling is required for remote actor lifecycle
management, it is better to ignore remote actors altogether and let them start
and stop on their own. With that, we rely on the correctness of desired status
and that client actors release all resources when they stop.

Also splitted the client actor state UNKNOWN into UNKNOWN and INITIALIZING.
The state UNKNOWN is for permanently failed clients, and the state INITIALIZING
has a short timeout to start remote client actors as soon as possible without
any signal from the connection persistent actor.

Signed-off-by: Yufei Cai <yufei.cai@bosch-si.com>