Enhance docs of docker deployment (fixes #443)
Add entrypoint section to compose file

To make it more obvious for the operator how a service may be
configured the entrypoint section used for the services is now
explicitly part of the compose file.

This way someone new to ditto and java has a better chance of
understanding how the services may be configured.

Add short summary on config to docker readme

To make it more clear this commit adds a section on configuration
to the docker deployment readme. This section handles two topics.

First the mechanics of a java application configuration are laid out.
The configuration is depicted via the entrypoint section of the
compose-file.
This approach is chosen, because it enables a fairly easy configuration
of the java service via the java VM parameters.
The chosen example is the devops command from the gateway service
due to it's practicability and importance when setting up an instance
for oneself.

Second follows a description on how to retrieve possible configuration
options either via HTTP REST API or via the given configuration files in
the repository. This should enable more people to faster learn how
to work with and configure each service.

Add a config section to operational docs

To enhance the understanding for configuration of the ditto
micro-services this commit adds a configuration section to the public docs
for operators. It contains a docker-specific example and hints on where
to find available configuration options.

This is necessary to give operators new to ditto and java applications
a better start on hosting and using their own instances of ditto.

Remove misleading configuration from sandbox

This will remove confusing and unecessary configuration from the sandbox
in the docker deployment.

Setting a devops user and password basic auth before the devops route is
not just uneccessary, because this is configured in the gateway service
already, it's also misleading to new operators of ditto.

This is an enhancement with regard to #443

Signed-off-by: Alexander Wellbrock <a.wellbrock@mailbox.org>