Extend concierge enforcement Contextual by askFuture.
Reason: To push query commands through the enforcement stream quickly
so that they do not block other commands from entering the stream.
Command order is only important for the first message of chained
ask-steps in the execution of a query command.

Also permitted parallel future scheduling inside an enforcement lane
and added unit test for it.

Signed-off-by: Yufei Cai <yufei.cai@bosch-si.com>